Separate mouse handler from feathergrid
